# Conflicts:
#	src/main/java/de/uniks/stp24/component/game/BuildingPropertiesComponent.java
#	src/main/java/de/uniks/stp24/controllers/GangCreationController.java
#	src/main/java/de/uniks/stp24/controllers/InGameController.java
#	src/test/java/de/uniks/stp24/AppTest2.java
#	src/test/java/de/uniks/stp24/appTestModules/InGameTestInitializer.java
#	src/test/java/de/uniks/stp24/appTestModules/InGameTestLoader.java
#	src/test/java/de/uniks/stp24/game/PauseMenuTest.java
#	src/test/java/de/uniks/stp24/game/islandOverview/IslandOverviewTestComponent.java
#	src/test/java/de/uniks/stp24/game/islandOverview/IslandOverviewTestInitializer.java

+++ src/main/java/de/uniks/stp24/dto/AggregateItemDto.java
@@ -2,7 +2,7 @@ package de.uniks.stp24.dto;
 
 public record AggregateItemDto(
         String variable,
-        int count,
-        int subtotal
+        double count,
+        double subtotal
 ) {
 }
